(GH-64) Add type support in pdk config RFC
Currently the pdk config RFC only gives examples for string type setting values.
There is no way to pass in, or tell PDK, that the datatype is not a string.

This commit adds the `--type` (aliased with `--as`) and `--force` flags, and
removes the redundant `--add` flag so that type hints can be given to PDK for
type conversion.